requirements.txt
mongoengine==0.10.5
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
itsdangerous==0.24
PyJWT==0.2.1
requests==2.3.0
requests-oauthlib==0.4.1
flask-mongoengine==0.7.5
